+++ dlib/image_transforms/interpolation.h
@@ -1718,7 +1718,7 @@ namespace dlib
         {
             if (rect.is_empty())
             {
-                cols = rows = static_cast<unsigned long>(std::round(std::sqrt((double)size)));
+                cols = rows = std::round(std::sqrt((double)size));
             }
             else
             {
