+++ dlib/lsh/hashes.h
@@ -73,7 +73,7 @@ namespace dlib
             const result_type& b
         ) const
         {
-            return static_cast<unsigned int>(hamming_distance(a,b));
+            return hamming_distance(a,b);
         }
 
     private:
@@ -112,10 +112,8 @@ namespace dlib
             const result_type& b
         ) const
         {
-            return static_cast<unsigned int>(
-                hamming_distance(a.first,b.first) + 
-                hamming_distance(a.second,b.second)
-            );
+            return hamming_distance(a.first,b.first) + 
+                   hamming_distance(a.second,b.second);
         }
 
     private:
